finish logic for articles whit comment and like

+++ src/App/Controllers/CommentController.php
@@ -26,31 +26,72 @@ class CommentController {
         $stmt->bindParam(':id_article',$commentaire->id_article);
         $stmt->bindParam(':id_reader',$commentaire->id_reader);
         $stmt->bindParam(':content',$commentaire->content);
+        return $stmt->execute();
     }
 
-    public function addcommentaire(){
-        if($_SERVER['REQUEST_METHOD']!== 'POST'){
-            header('Location: /');
-        };
-        $this->checkauth();
-        $commentaire = new Commentaire();
-        $commentaire->id_article = $_POST['id_article'];
-        $commentaire->id_reader = $_POST['id_reader'];
-        $commentaire->content = $_POST['content'];
+private function getCommentsByArticleId($id){
+    $conn = Database::getConnection();
+    $sql = "
+        SELECT c.*, u.first_name, u.last_name
+        FROM commentaires c
+        LEFT JOIN users u ON u.id = c.id_reader
+        WHERE c.id_article = ?
+        ORDER BY c.create_at DESC
+    ";
+    $stmt = $conn->prepare($sql);
+    $stmt->execute([$id]);
+    return $stmt->fetchAll(PDO::FETCH_ASSOC);
+}
 
-        if($this->create($commentaire)){
 
-        };
 
+
+    public function showcomments(){
+        $id = $_GET["id"];
+        $comments = $this->getCommentsByArticleId($id);
+        require_once __DIR__ . '/../Views/Comments/allComments.php';
     }
+
+public function addcommentaire() {
+    if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
+        header('Location: /');
+        exit();
+    }
+
+    $this->checkauth();
+
+    $id_article = $_POST['id_article'] ?? null;
+    $content    = $_POST['content'] ?? null;
+
+    if (!$id_article || !$content) {
+        die("cant add comment whit out value");
+    }
+
+    $commentaire = new Commentaire();
+    $commentaire->id_article = $id_article;
+    $commentaire->id_reader  = $_SESSION['user']['id'];
+    $commentaire->content    = $content;
+
+    if ($this->create($commentaire)) {
+        header('Location: /articles?id=' . $id_article); 
+        exit();
+    }
+}
+
+
 
     public function deletecommentaire(){
         $conn = Database::getconnection();
         if($_SERVER['REQUEST_METHOD']!== 'POST'){
-            header('Location: /');
+            header('Location: /articles');
+            exit();
         };
         $this->checkauth();
-        $id = $_POST['id'];
+        if(!isset($_POST['id_commentaire'])){
+            header('Location: /comment');
+            exit();
+        }
+        $id = $_POST['id_commentaire'];
         $id_reader = $_SESSION['user']['id'];
         $sql = "DELETE FROM {$this->table}
         WHERE id = ? AND id_reader = ?";
